Refactor user.rs for testability
