+++ tests/user_flow.rs
@@ -1,0 +1,2 @@
+#[tokio::test]
+async fn test_user_flow() {}
